requirements.txt

streamlit
opencv-python-headless
pillow
numpy
geocoder
